[VPCEP] Move `services` to common tags
